Move client/server into management package separately.
